+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Removals
+
+- Removes the `aria-label` from the `Form` component. The previous value of `"form"` was redundant and not required.
+
 ## 3.5.1 (December 19, 2024)
 
 ### Adds